bg_api/pyproject.toml
[project]
name = "bg-api"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "pingfury", email = "pingfury@outlook.com" }
]
dependencies = [
    "fastapi[standard]>=0.116.1",
    "uvicorn[standard]>=0.24.0",
    "openai>=1.0.0",
    "python-multipart>=0.0.6",
    "pillow>=10.0.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.0.0",
    "httpx>=0.25.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/bg_api"]
